Hello World!
This is `ex_teal_direct_upload`, a plugin for `ex_teal`. Teal is a new internal project from Motel.
Right now it's still a secret so shh.

The project was originally developed along side a client project and
so unfortunately most of the development history beyond this point is
only available to the members of Motel.